backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pypdf2==3.0.1
pdfplumber==0.11.0
python-dotenv==1.0.1
requests==2.31.0
redis==5.0.3
